requirements.txt
playwright==1.40.0
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
schedule==1.2.0
coloredlogs==15.0.1
discord-webhook==1.3.0
hcaptcha-challenger
